pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "argparse>=1.4.0",
    "jupyter>=1.1.1",
    "librosa>=0.11.0",
    "llvmlite>=0.44.0",
    "matplotlib>=3.10.1",
    "numba>=0.61.0",
    "numpy>=2.1",
    "pandas>=2.2.3",
    "polars-lts-cpu>=1.25.2",
    "ruff>=0.11.0",
    "scipy>=1.15.2",
    "soundfile>=0.13.1",
    "streamlit>=1.43.2",
    "watchdog>=6.0.0",
    "wave>=0.0.2",
]
